requirements.txt
google-genai
python-dotenv
Pillow
ffmpeg-python
google-generativeai
elevenlabs